Minor fixes.
darcs-hash:20070204123852-f1522-ac4c8045c86ebccd91c0e8d5d3dc9a04ba4551b8.gz

+++ src/filesystems/sleuthkit/python/sk.h
@@ -321,8 +321,8 @@ typedef struct {
 	uint32_t type;
 	uint32_t id;
     struct block *blocks;
-    long long readptr;
-    long long size;
+    uint64_t readptr;
+    uint64_t size;
 } skfile;
 
 static void skfile_dealloc(skfile *self);
